commentaries is working now

+++ src/servlets/NewsPost.java
@@ -1,5 +1,6 @@
 package servlets;
 
+import entities.Comment;
 import entities.News;
 import entities.User;
 import freemarker.template.Configuration;
@@ -15,6 +16,7 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.HashMap;
 
 public class NewsPost extends HttpServlet {
@@ -22,17 +24,25 @@ public class NewsPost extends HttpServlet {
         HttpSession session = request.getSession();
         String status = request.getParameter("two");
         int id = Integer.parseInt(request.getParameter("id"));
-        if(status.equals("Deny")){
-            Helper.getNewsService().deleteNews(id);
+        if (status != null) {
+            if (status.equals("Deny")) {
+                Helper.getNewsService().deleteNews(id);
+            } else {
+                Helper.getNewsService().acceptNews(id);
+            }
+            response.sendRedirect("/tcheck");
         } else {
-            Helper.getNewsService().acceptNews(id);
+            String user = (String) session.getAttribute("login");
+            String comment = (String)request.getParameter("comment");
+            Helper.getNewsService().comment(new Comment(user,new Date().toString(),comment,id));
+            response.sendRedirect("/post?id="+id);
         }
-        response.sendRedirect("/tcheck");
     }
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         response.setContentType("text/html");
         HttpSession session = request.getSession();
+        Helper.logged(request,response);
         String user = (String) session.getAttribute("login");
         try {
             int id = Integer.parseInt(request.getParameter("id"));
@@ -40,10 +50,13 @@ public class NewsPost extends HttpServlet {
             if (news.isAccepted()) {
                 //if (!Helper.logged(request,session,response)) {
                 Configuration cfg = Helper.getConfig(getServletContext());
-                Template tmpl = cfg.getTemplate("topic.html");
+                Template tmpl = cfg.getTemplate("topic.ftl");
                 HashMap<String, Object> root = new HashMap<>();
+                ArrayList<Comment> comments = Helper.getUserService().getComments(id);
                 root.put("form_url", request.getRequestURI());
                 root.put("news", news);
+                root.put("logged", user!=null);
+                root.put("comments",comments);
                 try {
                     tmpl.process(root, response.getWriter());
                 } catch (TemplateException e) {
@@ -60,7 +73,7 @@ public class NewsPost extends HttpServlet {
                     } else {
                         if (cuser.isAdmin()) {
                             Configuration cfg = Helper.getConfig(getServletContext());
-                            Template tmpl = cfg.getTemplate("new_topic_admin.html");
+                            Template tmpl = cfg.getTemplate("new_topic_admin.ftl");
                             HashMap<String, Object> root = new HashMap<>();
                             root.put("form_url", request.getRequestURI());
                             root.put("news", news);
